chore: more code, add fix for search-alias view
